Fixed broken quartus "command" and updated usage description
The quartus_wrapper script was erroring out due to lib problems. Modified
the .profile file to also add /usr/lib/x86_64-linux-gnu/libstdc++.so.6 to
LD_PRELOAD which resolved the issue. The paths added by .profile.addons were
also incorrect given we are using version 17.0, so I updated that as well.
Finally, I modified the README.md description to clarify that quartus_sh is
not an .sh file that the user needs to find and move, but rather part of the
Quartus Lite CLI command set.